+++ packages/jaspr/lib/src/server/server_app.dart
@@ -62,9 +62,6 @@ final class ServerApp {
     reloadClients();
 
     if (isFirstStartup) {
-      print('[INFO] Running server in ${kDebugMode ? 'debug' : 'release'} mode');
-      print('Serving at http://${kDebugMode ? 'localhost' : _server!.address.host}:${_server!.port}');
-
       if (kGenerateMode) {
         requestRouteGeneration('/');
       }