+++ packages/jaspr_cli/lib/src/commands/build_command.dart
@@ -8,16 +8,19 @@ import 'package:build_daemon/data/build_status.dart';
 import 'package:build_daemon/data/build_target.dart';
 import 'package:collection/collection.dart';
 import 'package:http/http.dart' as http show Client, Response;
+import 'package:io/ansi.dart';
 import 'package:path/path.dart' as p;
 
 import '../dev/util.dart';
 import '../helpers/css_helper.dart';
 import '../helpers/dart_define_helpers.dart';
 import '../helpers/flutter_helpers.dart';
+import '../helpers/print_logo.dart';
 import '../helpers/proxy_helper.dart';
 import '../logging.dart';
 import '../process_runner.dart';
 import '../project.dart';
+import '../utils.dart';
 import 'base_command.dart';
 
 class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
@@ -49,6 +52,13 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
       'target-arch',
       help: 'Compile to a specific target architecture (only in server mode)',
       allowed: ['arm', 'arm64', 'riscv64', 'x64'],
+    );
+    argParser.addOption(
+      'port',
+      abbr: 'p',
+      help:
+          'Specify a port to run the server on during static generation. '
+          'Defaults to {jaspr.port} from pubspec.yaml or "$defaultServePort".',
     );
     argParser.addFlag('experimental-wasm', help: 'Compile to wasm', negatable: false);
     argParser.addMultiOption(
@@ -122,9 +132,14 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
 
   @override
   Future<int> runCommand() async {
-    ensureInProject();
+    await ensureInProject();
+    printLogo();
 
-    logger.write('Building jaspr for ${project.requireMode.name} rendering mode.');
+    logger.write('Building ${cyan.wrap(project.name)} in ${cyan.wrap(project.requireMode.name)} rendering mode.');
+    if (!verbose) {
+      logger.write('Showing reduced log output. Pass --verbose to see all output.', level: Level.debug);
+    }
+    logger.write('\n');
 
     final dir = Directory('build/jaspr').absolute;
     final webDir = project.requireMode == JasprMode.server ? Directory('build/jaspr/web').absolute : dir;
@@ -174,7 +189,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
     final serverDefines = getServerDartDefines();
 
     if (project.requireMode == JasprMode.server) {
-      logger.write('Building server app...', progress: ProgressState.running);
+      logger.write('Building server app...', tag: Tag.server, progress: ProgressState.running);
 
       final String? targetOS = argResults!.option('target-os');
       final String? targetArch = argResults!.option('target-arch');
@@ -192,6 +207,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
       if (!['exe', 'aot-snapshot'].contains(compileTarget) && (targetOS != null || targetArch != null)) {
         logger.write(
           'Cross-compilation (--target-os or --target-arch) is only supported with --target=exe or --target=aot-snapshot. Current target: $compileTarget',
+          tag: Tag.server,
           level: Level.error,
         );
         return 1;
@@ -217,13 +233,13 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
 
       await watchProcess('server build', process, progress: 'Compiling server executable.');
     } else if (project.requireMode == JasprMode.static) {
-      logger.write('Preparing static rendering...');
+      logger.write('Preparing server for static rendering...', tag: Tag.server, progress: ProgressState.running);
 
       final Map<String, ({String? lastmod, String? changefreq, double? priority})?> generatedRoutes = {};
       final List<String> queuedRoutes = [];
 
       final serverStartedCompleter = Completer<void>();
-      final serverPort = project.port ?? defaultServePort;
+      final serverPort = argResults!.option('port') ?? project.port ?? defaultServePort;
 
       await startProxy(
         serverProxyPort,
@@ -274,18 +290,19 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
         'server',
         process,
         tag: Tag.server,
-        progress: 'Starting server app...',
+        progress: 'Starting server for static rendering...',
         onFail: () {
+          if (done) return false;
           logger.write('Server process failed unexpectedly.', level: Level.error, progress: ProgressState.completed);
-          return !done;
+          return true;
         },
       );
 
       await serverStartedCompleter.future;
 
-      logger.write('Server started', progress: ProgressState.completed);
+      logger.write('Server started', tag: Tag.server, progress: ProgressState.completed);
 
-      logger.write('Generating routes...');
+      logger.write('\nGenerating pages:');
 
       final httpClient = http.Client();
 
@@ -293,7 +310,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
         final route = queuedRoutes.removeLast();
 
         logger.write(
-          '(${generatedRoutes.length - queuedRoutes.length}/${generatedRoutes.length}) Generating route "$route" ...',
+          '${cyan.wrap('(${generatedRoutes.length - queuedRoutes.length}/${generatedRoutes.length})')} Generating "$route" ...',
         );
 
         http.Response response;
@@ -302,6 +319,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
         } catch (e) {
           logger.write(
             'Failed to generate route "$route". ($e)',
+            tag: Tag.cli,
             level: Level.error,
           );
           hasBuildError = true;
@@ -311,6 +329,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
         if (response.statusCode != 200) {
           logger.write(
             'Failed to generate route "$route". (Received status code ${response.statusCode})',
+            tag: Tag.cli,
             level: Level.error,
           );
           hasBuildError = true;
@@ -337,6 +356,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
             if (sitemap is! Map<String, Object?>) {
               logger.write(
                 'Invalid sitemap data for route "$route". Expected a map, but got ${sitemap.runtimeType}.',
+                tag: Tag.cli,
                 level: Level.error,
               );
               hasBuildError = true;
@@ -364,7 +384,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
       logger.complete(!hasBuildError);
 
       if (sitemapDomain != null) {
-        logger.write('Generating sitemap.xml...');
+        logger.write('Generating sitemap.xml...', tag: Tag.cli);
 
         final content = StringBuffer();
         content.writeln('<?xml version="1.0" encoding="UTF-8"?>');
@@ -394,7 +414,9 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
           if (sitemapData.changefreq != null) {
             content.writeln('    <changefreq>${sitemapData.changefreq}</changefreq>');
           }
-          content.writeln('    <priority>${sitemapData.priority ?? 0.5}</priority>');
+          if (sitemapData.priority != null) {
+            content.writeln('    <priority>${sitemapData.priority}</priority>');
+          }
           content.writeln('  </url>');
         }
         content.writeln('</urlset>');
@@ -419,11 +441,17 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
     }
 
     if (hasBuildError) {
-      logger.write('Failed to build project. Check the logs above for more information.', level: Level.error);
+      logger.write(
+        'Failed to build ${cyan.wrap(project.name)}. Check the logs above for more information.',
+        level: Level.error,
+      );
       return 1;
     }
 
-    logger.write('Completed building project to /build/jaspr.');
+    logger.write('');
+    logger.write(
+      wrapBox('Completed building ${cyan.wrap(project.name)} to ${cyan.wrap('build/jaspr/')}', borderColor: green),
+    );
     return 0;
   }
 
@@ -432,7 +460,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
       project.checkWasmSupport();
     }
 
-    logger.write('Building web assets...', progress: ProgressState.running);
+    logger.write('Building web assets...', tag: Tag.builder, progress: ProgressState.running);
 
     final compiler = useWasm ? 'dart2wasm' : 'dart2js';
 
@@ -539,7 +567,7 @@ class BuildCommand extends BaseCommand with ProxyHelper, FlutterHelper {
     }
     await client.close();
     if (exitCode == 0) {
-      logger.write('Completed building web assets.', progress: ProgressState.completed);
+      logger.write('Completed building web assets.', tag: Tag.builder, progress: ProgressState.completed);
 
       if (includeSourceMaps) {
         _fixSourceMaps(outputLocation.output);