+++ packages/jaspr_cli/test/integration/commands_integration_test.dart
@@ -128,6 +128,6 @@ class TestRunner {
   }
 
   void setup() {
-    runner = JasprCommandRunner();
+    runner = JasprCommandRunner(true, false);
   }
 }