+++ packages/jaspr_cli/test/src/commands/build_command_test.dart
@@ -18,7 +18,7 @@ void main() {
 
     setUp(() {
       io = FakeIO();
-      runner = JasprCommandRunner(false);
+      runner = JasprCommandRunner(false, false);
     });
 
     tearDown(() {
@@ -34,11 +34,11 @@ void main() {
 
         final buildResult = runner.run(['build', '--verbose']);
 
-        await expectLater(io.stdout.queue, emits('Building jaspr for client rendering mode.'));
+        await expectLater(io.stdout.queue, emits('Building myapp in client rendering mode.'));
 
         await io.runReleaseBuild(buildDaemon);
 
-        await expectLater(io.stdout.queue, emits('Completed building project to /build/jaspr.'));
+        await expectLater(io.stdout.queue, emitsThrough(contains('Completed building')));
 
         expect(await buildResult, equals(0));
       });
@@ -71,10 +71,7 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            'Building jaspr for server rendering mode.',
-            'Using server entry point: lib/main.server.dart',
-          ]),
+          emits('Building myapp in server rendering mode.'),
         );
 
         await io.runReleaseBuild(buildDaemon);
@@ -82,14 +79,14 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Building server app...',
+            '[SERVER] Building server app...',
             'Compiling server executable.',
           ]),
         );
 
         serverProcess.exit(0);
 
-        await expectLater(io.stdout.queue, emits('Completed building project to /build/jaspr.'));
+        await expectLater(io.stdout.queue, emitsThrough(contains('Completed building')));
 
         expect(await buildResult, equals(0));
       });
@@ -133,10 +130,7 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            'Building jaspr for static rendering mode.',
-            'Using server entry point: lib/main.server.dart',
-          ]),
+          emits('Building myapp in static rendering mode.'),
         );
 
         await io.runReleaseBuild(buildDaemon);
@@ -144,8 +138,8 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Preparing static rendering...',
-            '[SERVER] Starting server app...',
+            '[SERVER] Preparing server for static rendering...',
+            '[SERVER] Starting server for static rendering...',
           ]),
         );
 
@@ -165,9 +159,10 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Server started',
-            'Generating routes...',
-            '(1/1) Generating route "/abc" ...',
+            '[SERVER] Server started',
+            '',
+            'Generating pages:',
+            '(1/1) Generating "/abc" ...',
           ]),
         );
 
@@ -177,8 +172,87 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Completed building project to /build/jaspr.',
-            'Terminating server...',
+            emitsThrough(contains('Completed building')),
+            emitsThrough('Terminating server...'),
+          ]),
+        );
+
+        expect(await buildResult, equals(0));
+      });
+    });
+
+    test('builds static project with custom port', () async {
+      await io.runZoned(() async {
+        io.setupFakeProject('myapp', mode: 'static');
+        io.stubDartSDK();
+
+        final buildDaemon = io.setupFakeBuildDaemon(verifyArgs: buildRunnerBuildArgs);
+
+        FakeProcess? serverProcess;
+        when(
+          () => io.process.start(
+            '/fake/bin/dart',
+            any(that: contains('-Djaspr.flags.generate=true')),
+            workingDirectory: '/root/myapp',
+            environment: any(named: 'environment'),
+          ),
+        ).thenAnswer((inv) async {
+          expect(
+            inv.namedArguments[#environment],
+            equals({'PORT': '8181', 'JASPR_PROXY_PORT': serverProxyPort}),
+          );
+
+          return serverProcess = FakeProcess();
+        });
+
+        final buildResult = runner.run(['build', '--port', '8181', '--verbose']);
+
+        await expectLater(
+          io.stdout.queue,
+          emits('Building myapp in static rendering mode.'),
+        );
+
+        await io.runReleaseBuild(buildDaemon);
+
+        await expectLater(
+          io.stdout.queue,
+          emitsInOrder([
+            '[SERVER] Preparing server for static rendering...',
+            '[SERVER] Starting server for static rendering...',
+          ]),
+        );
+
+        expect(serverProcess, isNotNull);
+
+        final proxyConnection = await io.connectToProxy();
+
+        final fakeServer = io.setupFakeServer(8181);
+
+        fakeServer.onRequest('/abc', (request) async {
+          request.response.write('FAKE HTML RESPONSE');
+          await request.response.close();
+        });
+
+        await proxyConnection.sendFakeRequest(r'$jasprMessageHandler', jsonEncode({'route': '/abc'}));
+
+        await expectLater(
+          io.stdout.queue,
+          emitsInOrder([
+            '[SERVER] Server started',
+            '',
+            'Generating pages:',
+            '(1/1) Generating "/abc" ...',
+          ]),
+        );
+
+        expect(io.fs.file('/root/myapp/build/jaspr/abc/index.html').existsSync(), isTrue);
+        expect(io.fs.file('/root/myapp/build/jaspr/abc/index.html').readAsStringSync(), equals('FAKE HTML RESPONSE'));
+
+        await expectLater(
+          io.stdout.queue,
+          emitsInOrder([
+            emitsThrough(contains('Completed building')),
+            emitsThrough('Terminating server...'),
           ]),
         );
 
@@ -196,7 +270,7 @@ void main() {
 
         final buildResult = runner.run(['build', '--verbose']);
 
-        await expectLater(io.stdout.queue, emits('Building jaspr for client rendering mode.'));
+        await expectLater(io.stdout.queue, emits('Building myapp in client rendering mode.'));
 
         expect(io.fs.file('/root/myapp/web/index.html').existsSync(), isTrue);
 
@@ -226,7 +300,7 @@ void main() {
 
         await io.runReleaseBuild(buildDaemon);
 
-        await expectLater(io.stdout.queue, emits('Completed building project to /build/jaspr.'));
+        await expectLater(io.stdout.queue, emitsThrough(contains('Completed building')));
 
         expect(io.fs.file('/root/myapp/build/jaspr/flutter_bootstrap.js').existsSync(), isTrue);
         expect(
@@ -268,10 +342,7 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            'Building jaspr for static rendering mode.',
-            'Using server entry point: lib/main.server.dart',
-          ]),
+          emits('Building myapp in static rendering mode.'),
         );
 
         await io.runReleaseBuild(buildDaemon);
@@ -279,8 +350,8 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Preparing static rendering...',
-            '[SERVER] Starting server app...',
+            '[SERVER] Preparing server for static rendering...',
+            '[SERVER] Starting server for static rendering...',
           ]),
         );
 
@@ -300,7 +371,6 @@ void main() {
             jsonEncode({
               'lastmod': '2026-01-01T12:34:56.789Z',
               'changefreq': 'daily',
-              'priority': 0.3,
             }),
           );
           request.response.write('FAKE HTML RESPONSE 2');
@@ -316,11 +386,12 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Server started',
-            'Generating routes...',
-            '(1/1) Generating route "/abc" ...',
-            '(2/2) Generating route "/abc2" ...',
-            'Generating sitemap.xml...',
+            '[SERVER] Server started',
+            '',
+            'Generating pages:',
+            '(1/1) Generating "/abc" ...',
+            '(2/2) Generating "/abc2" ...',
+            '[LOG] Generating sitemap.xml...',
           ]),
         );
 
@@ -349,7 +420,6 @@ void main() {
             '    <loc>https://example.com/abc2</loc>\n'
             '    <lastmod>2026-01-01T12:34:56Z</lastmod>\n'
             '    <changefreq>daily</changefreq>\n'
-            '    <priority>0.3</priority>\n'
             '  </url>\n'
             '</urlset>\n'
             '',
@@ -359,8 +429,8 @@ void main() {
         await expectLater(
           io.stdout.queue,
           emitsInOrder([
-            'Completed building project to /build/jaspr.',
-            'Terminating server...',
+            emitsThrough(contains('Completed building')),
+            emitsThrough('Terminating server...'),
           ]),
         );
 
@@ -377,7 +447,7 @@ void main() {
 
         final buildResult = runner.run(['build', '--verbose']);
 
-        await expectLater(io.stdout.queue, emits('Building jaspr for client rendering mode.'));
+        await expectLater(io.stdout.queue, emits('Building myapp in client rendering mode.'));
 
         final runnerFilePath = '.dart_tool/build/generated/myapp/lib/main.client.styles.dart';
         io.fs.file('/root/myapp/$runnerFilePath')
@@ -400,7 +470,11 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder(['[CLI] Generated main.css', 'Completed building project to /build/jaspr.']),
+          emitsInOrder([
+            '[STYLES] Generating CSS...',
+            '[STYLES] Generated main.css',
+            emitsThrough(contains('Completed building')),
+          ]),
         );
 
         final tempOutputPath = '/root/myapp/.dart_tool/jaspr/generated/main.css';